pyproject.toml
[project]
name = "reviewer-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.42.42",
    "chromadb>=1.4.1",
    "docx2txt>=0.9",
    "faiss-cpu>=1.13.2",
    "ipython>=9.10.0",
    "json-repair>=0.56.0",
    "langchain>=1.2.8",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.8",
    "langchain-google-genai>=4.2.0",
    "langchain-google-vertexai>=3.2.2",
    "langchain-groq>=1.1.2",
    "langchain-huggingface>=1.2.0",
    "langchain-openai>=1.1.7",
    "langchain-postgres>=0.0.16",
    "langgraph>=1.0.7",
    "phidata>=2.7.10",
    "psycopg2-binary>=2.9.11",
    "pymongo>=4.16.0",
    "pymupdf>=1.26.7",
    "pypdf>=6.6.2",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.2.2",
    "typesense>=1.3.0",
]
